Angle-side-side (or side-side-angle)
The most interesting of the triangle configurations in plane geometry as this one can have two distinct solutions up to congruence.
This completes the illustrations of triangle constructions given three of the six defining components (sides and interior) of a triangle in the plane.  The missing configuration (AAA) yields an infinitude of non-congruent solutions in the plane.